backend-python/requirements.txt
# Flask web framework
Flask>=3.0.0
flask-cors>=4.0.0

# Database
mysql-connector-python>=8.2.0

# Document parsing
python-docx>=1.1.0

# Vector database
qdrant-client>=1.7.0

# AI/ML
openai>=1.6.0
anthropic>=0.8.0

# Environment management
python-dotenv>=1.0.0

# WSGI server (for local development)
gunicorn>=21.2.0

# Werkzeug (Flask dependency, pinned for compatibility)
Werkzeug>=3.0.0
